+++ plugins/auth-backend/src/providers/aws-alb/provider.ts
@@ -212,28 +212,6 @@ export class AwsAlbAuthProvider implements AuthProviderRouteHandlers {
 }
 
 /**
- * @public
- * @deprecated This type has been inlined into the create method and will be removed.
- */
-export type AwsAlbProviderOptions = {
-  /**
-   * The profile transformation function used to verify and convert the auth response
-   * into the profile that will be presented to the user.
-   */
-  authHandler?: AuthHandler<AwsAlbResult>;
-
-  /**
-   * Configure sign-in for this provider, without it the provider can not be used to sign users in.
-   */
-  signIn: {
-    /**
-     * Maps an auth result to a Backstage identity for the user.
-     */
-    resolver: SignInResolver<AwsAlbResult>;
-  };
-};
-
-/**
  * Auth provider integration for AWS ALB auth
  *
  * @public
@@ -282,9 +260,3 @@ export const awsAlb = createAuthProviderIntegration({
     };
   },
 });
-
-/**
- * @public
- * @deprecated Use `providers.awsAlb.create` instead
- */
-export const createAwsAlbProvider = awsAlb.create;