+++ plugins/auth-backend/src/providers/bitbucket/provider.ts
@@ -191,28 +191,6 @@ export class BitbucketAuthProvider implements OAuthHandlers {
 }
 
 /**
- * @public
- * @deprecated This type has been inlined into the create method and will be removed.
- */
-export type BitbucketProviderOptions = {
-  /**
-   * The profile transformation function used to verify and convert the auth response
-   * into the profile that will be presented to the user.
-   */
-  authHandler?: AuthHandler<OAuthResult>;
-
-  /**
-   * Configure sign-in for this provider, without it the provider can not be used to sign users in.
-   */
-  signIn?: {
-    /**
-     * Maps an auth result to a Backstage identity for the user.
-     */
-    resolver: SignInResolver<OAuthResult>;
-  };
-};
-
-/**
  * Auth provider integration for BitBucket auth
  *
  * @public
@@ -305,23 +283,3 @@ export const bitbucket = createAuthProviderIntegration({
     },
   },
 });
-
-/**
- * @public
- * @deprecated Use `providers.bitbucket.create` instead
- */
-export const createBitbucketProvider = bitbucket.create;
-
-/**
- * @public
- * @deprecated Use `providers.bitbucket.resolvers.usernameMatchingUserEntityAnnotation()` instead.
- */
-export const bitbucketUsernameSignInResolver =
-  bitbucket.resolvers.usernameMatchingUserEntityAnnotation();
-
-/**
- * @public
- * @deprecated Use `providers.bitbucket.resolvers.userIdMatchingUserEntityAnnotation()` instead.
- */
-export const bitbucketUserIdSignInResolver =
-  bitbucket.resolvers.userIdMatchingUserEntityAnnotation();