+++ plugins/adr-common/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-adr-common
 
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.0-next.0
+  - @backstage/integration@1.2.2-next.0
+
 ## 0.1.1
 
 ### Patch Changes