+++ plugins/auth-node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-auth-node
 
+## 0.2.3-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.0
+
 ## 0.2.2
 
 ### Patch Changes