+++ plugins/badges-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-badges-backend
 
+## 0.1.28-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.0
+  - @backstage/catalog-model@1.1.0-next.0
+  - @backstage/catalog-client@1.0.4-next.0
+
 ## 0.1.27
 
 ### Patch Changes