+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-bazaar
 
+## 0.1.22-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-model@1.1.0-next.0
+  - @backstage/core-components@0.9.6-next.0
+  - @backstage/plugin-catalog@1.3.1-next.0
+  - @backstage/cli@0.17.3-next.0
+  - @backstage/catalog-client@1.0.4-next.0
+  - @backstage/plugin-catalog-react@1.1.2-next.0
+
 ## 0.1.21
 
 ### Patch Changes