+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-catalog-import
 
+## 0.8.10-next.0
+
+### Patch Changes
+
+- 272106fdad: Support use without `integrations` or only integrations without frontend visible properties (e.g., `bitbucketCloud`) being configured by checking `integrations.github` directly without attempting to load `integrations`.
+- Updated dependencies
+  - @backstage/catalog-model@1.1.0-next.0
+  - @backstage/core-components@0.9.6-next.0
+  - @backstage/integration@1.2.2-next.0
+  - @backstage/catalog-client@1.0.4-next.0
+  - @backstage/plugin-catalog-react@1.1.2-next.0
+  - @backstage/integration-react@1.1.2-next.0
+
 ## 0.8.9
 
 ### Patch Changes