+++ plugins/periskop-backend/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @backstage/plugin-periskop-backend
 
+## 0.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.0
+
 ## 0.1.4
 
 ### Patch Changes