+++ plugins/techdocs-module-addons-contrib/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-techdocs-module-addons-contrib
 
+## 1.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.6-next.0
+  - @backstage/plugin-techdocs-react@1.0.2-next.0
+  - @backstage/integration@1.2.2-next.0
+  - @backstage/integration-react@1.1.2-next.0
+
 ## 1.0.1
 
 ### Patch Changes