+++ plugins/vault-backend/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-vault-backend
 
+## 0.2.0-next.0
+
+### Minor Changes
+
+- 5ebf2c7023: Throw exceptions instead of swallow them, remove some exported types from the `api-report`, small changes in the API responses & expose the vault `baseUrl` to the frontend as well
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.0
+  - @backstage/backend-tasks@0.3.3-next.0
+  - @backstage/backend-test-utils@0.1.26-next.0
+
 ## 0.1.0
 
 ### Minor Changes